--- a/examples/text-generation/run_generation.py
+++ b/examples/text-generation/run_generation.py
@@ -606,6 +606,11 @@ def main():
                     max_length=args.max_input_tokens,
                     truncation=True,
                 )
+                def compute_valid_sequence_lengths_tensor(input_tokens):
+                    attn_mask = input_tokens['attention_mask']
+                    return torch.sum(attn_mask, dim=1)
+                valid_sequence_lengths = compute_valid_sequence_lengths_tensor(input_tokens).to(args.device)
+                setattr(generation_config, 'valid_sequence_lengths', valid_sequence_lengths)
             else:
                 input_tokens = tokenizer.batch_encode_plus(input_sentences, return_tensors="pt", padding=True)
             encode_duration = time.perf_counter() - encode_t0