Valid sequence length (VSL) feature provides a way of indicating to the FusedSDPA operator which part of inputs contain valid information and which are just padding. Thanks to that operator can skip unnecessary computation which improves performance.

This change also enables using "causal" flag for the prefill in case inputs are of different length. Previously using causal with different length of inputs would cause padding to be included in computation which caused degradation in accuracy.


Below is performance data showcasing the improvement in the model with causal flag:

BS - batch size
MIT - max input tokens
MNT - max new tokens
VSL - valid sequence lengths

BF16:

BS | MIT | MNT | Causal - Lower Triangular Softmax (Through VSL Feature) | Throughput (Including Tokenization) Tokens/s | Improvement vs. Without Causal [%]
-- | -- | -- | -- | -- | --
12 | 31744 | 1024 | 0 | 104.79 |  
12 | 31744 | 1024 | 1 | 182.095 | 53.89
16 | 24576 | 8192 | 0 | 401.74 |  
16 | 24576 | 8192 | 1 | 455.85 | 12.62
24 | 16384 | 16384 | 0 | 623.25 |  
24 | 16384 | 16384 | 1 | 649.78 | 4.17
48 | 8192 | 8192 | 0 | 1242.62 |  
48 | 8192 | 8192 | 1 | 1285.92 | 3.42
96 | 4096 | 4096 | 0 | 2368.83 |

FP8:

BS | MIT | MNT | Causal - Lower Triangular Softmax (Through VSL Feature) | Throughput (including   tokenization) tokens/s | Improvement vs without causal [%]
-- | -- | -- | -- | -- | --
1230 | 128 | 128 | 0 | 13400.8 |  
1230 | 128 | 128 | 1 | 13652.56 | 1.86
225 | 2048 | 2048 | 0 | 5244.52 |  
225 | 2048 | 2048 | 1 | 5384.09 | 2.63
114 | 4096 | 4096 | 0 | 3441.98 |  
114 | 4096 | 4096 | 1 | 3507.86 | 1.90
37 | 8192 | 8192 | 0 | 1261.14 |  
37 | 8192 | 8192 | 1 | 1321.64 | 4.68